requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
lxml==6.0.2
numpy==2.2.6
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
